api-store/requirements.txt
alembic==1.9.2
fastapi==0.89.1
psycopg2-binary==2.9.5
pydantic==1.10.4
SQLAlchemy==2.0.1
uvicorn==0.20.0
pytest==7.2.1
factory-boy==3.2.1
Faker==16.8.1
python-dotenv
